requirements-file.txt
flask==2.3.3
flask-restful==0.3.10
langchain==0.1.1
langchain-community==0.0.13
langchain-google-genai==0.0.6
faiss-cpu==1.7.4
google-generativeai==0.3.1
python-dotenv==1.0.0
